shell: Do not attempt always bind to agl-shell interface
This helps up running as a regular application in
situations where we already have a client bound to the agl-shell
interface.

We can't really accept more than one client to bind to the interface and
without a change in the protocol to avoid racy situations, it seems more
natural for the embedder to infer if it actually needs to do that.

When the protocol has been updated to include additional events to let
clients that they've lost the race, we can remove this change.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>