agl-shell: Add support for v4 agl-shell protocol
This adds the ability the specify the activation area programmatically
rather than doing it from the ini configuration file.

This also includes an explicit activation area of the 0, 160 - which
set-ups an rectangle that allows two panels up and down each of 160
width in size, to match the explicit activation area we have already in
our configuration ini. A further patch would allow to customize that
using the json configuration files.

Bug-AGL: SPEC-4603, SPEC-4784
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>